fix: use u32 for milliseconds since startup

+++ components/util/util.cpp
@@ -1,3 +1,3 @@
 #include "util.h"
 
-ulong millis() { return esp_timer_get_time() / 1000; }
+u32 millis() { return esp_timer_get_time() / 1000; }
